move logging stuff to logging module
